+++ app/src/main/java/com/example/superhero_room/Activities/MainActivity.kt
@@ -1,7 +1,11 @@
 package com.example.superhero_room.Activities
 
+import android.graphics.Bitmap
+import android.graphics.BitmapFactory
+import android.graphics.drawable.Drawable
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.core.graphics.drawable.toBitmap
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.StaggeredGridLayoutManager
@@ -13,6 +17,9 @@ import com.example.superhero_room.databinding.ActivityMainBinding
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
+import java.io.ByteArrayInputStream
+import java.io.ByteArrayOutputStream
+import java.nio.ByteBuffer
 
 class MainActivity : AppCompatActivity() {
     lateinit var binding: ActivityMainBinding
@@ -25,11 +32,12 @@ class MainActivity : AppCompatActivity() {
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        var superheros = arrayListOf("Batman", "Superman", "Wonder Woman")
+        var superheros = arrayListOf("Batman", "Superman", "Wonder Woman", "Flash")
         var images = arrayListOf(
             getDrawable(R.drawable.batman),
             getDrawable(R.drawable.superman),
-            getDrawable(R.drawable.wonderwoman)
+            getDrawable(R.drawable.wonderwoman),
+            getDrawable(R.drawable.flash)
         )
 
         viewModel = ViewModelProvider(this)[MainActivityViewModel::class.java]
@@ -39,19 +47,23 @@ class MainActivity : AppCompatActivity() {
             StaggeredGridLayoutManager(2, StaggeredGridLayoutManager.VERTICAL)
 
         coroutineScope.launch {
-            //   addToDatabase()
+            addToDatabase()
+            val bitmapBatman = BitmapFactory.decodeResource(resources, R.drawable.batman)
+            val bitmapSuperman = BitmapFactory.decodeResource(resources, R.drawable.superman)
+            val bitmapWonderWoman = BitmapFactory.decodeResource(resources, R.drawable.wonderwoman)
+            val bitmapFlash = BitmapFactory.decodeResource(resources, R.drawable.flash)
+
+            addImagesToInternalStorage("Superman.jpg", bitmapSuperman);
+            addImagesToInternalStorage("Batman.jpg", bitmapBatman)
+            addImagesToInternalStorage("Wonder Woman.jpg", bitmapWonderWoman)
+            addImagesToInternalStorage("Flash.jpg", bitmapFlash)
         }
 
     }
 
     private suspend fun addToDatabase() {
         val supermanEntity: SuperheroEntity = SuperheroEntity(
-            0,
-            "Superman",
-            "This is superman",
-            "190 cm",
-            "106 kg",
-            "Kryptonian",
+            0, "Superman", getString(R.string.superman_description), "190 cm", "106 kg",
             /*  arrayOf(
                   "Healing",
                   "Electricity resistance",
@@ -59,12 +71,63 @@ class MainActivity : AppCompatActivity() {
                   "Laser eyes",
                   "Flight"
               ),*/
-            "90",
-            "100",
-            "100",
-            "100",
-            "70", "random"
+            90F, 100F, 100F, 100F, 100F, "random"
         )
         viewModel.insertSuperHero(supermanEntity)
+
+        val batmanEntity = SuperheroEntity(
+            0,
+            "Batman",
+            getString(R.string.batman_description),
+            "188 cm",
+            "95 kg",
+            100F,
+            40F,
+            29F,
+            19F,
+            100F,
+            "random"
+        )
+        viewModel.insertSuperHero(batmanEntity)
+
+        val wonderWomanEntity = SuperheroEntity(
+            0,
+            "Wonder Woman",
+            getString(R.string.wonderWoman_description),
+            "185 cm",
+            "77 kgs",
+            75F,
+            90F,
+            85F,
+            100F,
+            65F,
+            "random"
+        )
+
+        viewModel.insertSuperHero(wonderWomanEntity)
+
+        val flashEntity = SuperheroEntity(
+            0,
+            "Flash",
+            getString(R.string.flash_description),
+            "184 cm",
+            "90 kgs",
+            100F,
+            85F,
+            100F,
+            80F,
+            80F,
+            "random"
+        )
+        viewModel.insertSuperHero(flashEntity)
+    }
+
+    private fun addImagesToInternalStorage(name: String, bitmap: Bitmap) {
+        openFileOutput(name, MODE_PRIVATE).use {
+            var byteArrayOutputStream = ByteArrayOutputStream()
+            bitmap.compress(Bitmap.CompressFormat.JPEG, 100, byteArrayOutputStream)
+            var byteArray = byteArrayOutputStream.toByteArray()
+            it.write(byteArray)
+        }
     }
 }