+++ ChangeLog.md
@@ -2,10 +2,15 @@ CHANGELOG
 --------------------------------
 --------------------------------
 
+ver 1.2.2(23rd March 2023)
+-------------------------------
+--> bar chart added
+--> Loading of images from internal storage works now
+--> Ready for data entry
+
 ver 1.1.1 (14th March 2023)
 --------------------------------
 --> Detailed Activity basic UI
-
 
 ver 1.1.0 (13th March 2023)
 --------------------------------